Move ScanAnNft into its own component
