+++ src/routes/index.svelte
@@ -1,10 +1,5 @@
 <script lang="ts">
-	let openseaUrl: string;
-
-	async function evaluateNft(openseaUrl: string): Promise<void> {
-		console.log("boop");
-
-	}
+	import ScanAnNft from "$lib/components/ScanAnNft.svelte";
 </script>
 
 <svelte:head>
@@ -36,41 +31,7 @@
 			<svg xmlns="http://www.w3.org/2000/svg" class="text-white h-16 w-16" fill="none" viewBox="0 0 24 24" stroke="currentColor"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 11H5m14 0a2 2 0 012 2v6a2 2 0 01-2 2H5a2 2 0 01-2-2v-6a2 2 0 012-2m14 0V9a2 2 0 00-2-2M5 11V9a2 2 0 012-2m0 0V5a2 2 0 012-2h6a2 2 0 012 2v2M7 7h10"></path></svg>
 		</div>
 
-		<!-- todo make component -->
-		<form on:submit|preventDefault={() => evaluateNft('hi', 'hi') }>
-			<div class='mb-4'>
-				<p class='mb-1'>OpenSea URL</p>
-				<input
-					class='border border-gray-400 rounded px-2 py-1 w-full'
-					value={openseaUrl}
-				/>
-			</div>
-
-			<div class='border border-t-1 border-gray-400 my-6'></div>
-
-			<div class='mb-4'>
-				<p class='mb-1'>NFT Contract Address</p>
-				<input
-					class='border border-gray-400 rounded px-2 py-1 w-full'
-					value={openseaUrl}
-				/>
-			</div>
-
-			<div class='mb-4'>
-				<p class='mb-1'>Token ID</p>
-				<input
-					class='border border-gray-400 rounded px-2 py-1 w-full'
-					value={openseaUrl}
-				/>
-			</div>
-
-			<input
-				class:disabled={true}
-				class='w-full bg-gray-900 rounded-full text-white px-8 py-4 border border-white mb-8'
-				type="submit"
-				value="Analyze NFT"
-			/>
-		</form>
+		<ScanAnNft />
 	</div>
 
 	<div class='p-4 mt-12'>
